python/requirements.txt
Cython==0.20.1
biopython==1.63
numpy==1.8.0
pysam==0.7.7
scipy==0.13.3
seqmagick==0.5.0
